WINTERMUTE: Explicit casting in transpaprent_surface.cpp

+++ engines/wintermute/graphics/transparent_surface.cpp
@@ -70,25 +70,25 @@ void TransparentSurface::copyPixelBilinear(float projX, float projY, int dstX, i
 			if (x1 >= srcW || x1 < 0 || y1 >= srcH || y1 < 0) { 
 				Q11 = 0;
 			} else {
-				Q11 = READ_UINT32((const byte *)src->getBasePtr(x1 + srcRect.left, y1 + srcRect.top));
+				Q11 = READ_UINT32((const byte *)src->getBasePtr((int)(x1 + srcRect.left),(int)(y1 + srcRect.top)));
 			}
 
 			if (x1 >= srcW || x1 < 0 || y2 >= srcH || y2 < 0) { 
 				Q12 = 0;
 			} else {
-				Q12 = READ_UINT32((const byte *)src->getBasePtr(x1 + srcRect.left, y2 + srcRect.top));
+				Q12 = READ_UINT32((const byte *)src->getBasePtr((int)(x1 + srcRect.left), (int)(y2 + srcRect.top)));
 			}
 
 			if (x2 >= srcW || x2 < 0 || y1 >= srcH || y1 < 0) { 
 				Q21 = 0;
 			} else {
-				Q21 = READ_UINT32((const byte *)src->getBasePtr(x2 + srcRect.left, y1 + srcRect.top));
+				Q21 = READ_UINT32((const byte *)src->getBasePtr((int)(x2 + srcRect.left), (int)(y1 + srcRect.top)));
 			}
 
 			if (x2 >= srcW || x2 < 0 || y2 >= srcH || y2 < 0) { 
 				Q22 = 0;
 			} else {
-				Q22 = READ_UINT32((const byte *)src->getBasePtr(x2 + srcRect.left, y2 + srcRect.top));
+				Q22 = READ_UINT32((const byte *)src->getBasePtr((int)(x2 + srcRect.left), (int)(y2 + srcRect.top)));
 			}
 
 			byte *Q11s = (byte *)&Q11;
@@ -579,8 +579,8 @@ TransparentSurface *TransparentSurface::scale(uint16 newWidth, uint16 newHeight)
 	int projY;
 	for (int y = 0; y < dstH; y++) {
 		for (int x = 0; x < dstW; x++) {
-			projX = x / (float)dstW * srcW;
-			projY = y / (float)dstH * srcH;
+			projX = (int)(x / (float)dstW * srcW);
+			projY = (int)(y / (float)dstH * srcH);
 			copyPixelNearestNeighbor(projX, projY, x, y, srcRect, dstRect, this, target); 
 		}
 	}
